OrgHyperlinkInsert can insert relative links
OrgHyperlinkInsert (gil) will now try to make the inserted
link relative to the current file's directory. Utl resolves a relative
path based on the directory of the file containing the link, so these
relative links can be followed. This behavior matches org-mode.

A new command, OrgHyperlinkInsertAbsolute, has been added which inserts
the link using the old behavior (no attempt is made to make the link
relative). This has been bound to gIL.